+++ cache2/src/io/engine/posix.rs
@@ -174,6 +174,13 @@ impl BackendIoEngine {
 }
 
 impl IoEngine for BackendIoEngine {
+    fn set_latency_recorder(&self, recorder: crate::stats::recording::IoTiming) {
+        assert!(
+            self.inner.shared.latency.set(recorder).is_ok(),
+            "I/O recorder installed twice"
+        );
+    }
+
     fn try_reserve_read(&self) -> io::Result<ReadSlot> {
         self.inner.try_reserve_read()
     }